# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Thinking Spree consultant platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sprints:
    get:
      operationId: listSprints
      tags: [sprints]
      summary: List T-Sprint sessions (scoped to caller by default)
      description: >
        Default (scope=mine) returns only sprints aligned to the caller —
        as consultant, host, or co-host. Pass scope=all to get the team-wide
        register used by the Sprint Tracking page.
      parameters:
        - name: scope
          in: query
          required: false
          schema:
            type: string
            enum: [mine, all]
            default: mine
      responses:
        "200":
          description: List of sprints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sprint"
    post:
      operationId: createSprint
      tags: [sprints]
      summary: Create a new T-Sprint session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SprintInput"
      responses:
        "201":
          description: Created sprint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sprint"
  /sprints/{id}:
    get:
      operationId: getSprint
      tags: [sprints]
      summary: Get a specific T-Sprint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sprint details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sprint"
        "404":
          description: Not found
    patch:
      operationId: updateSprint
      tags: [sprints]
      summary: Update a T-Sprint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SprintUpdate"
      responses:
        "200":
          description: Updated sprint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sprint"
    delete:
      operationId: deleteSprint
      tags: [sprints]
      summary: Delete a sprint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /sprints/{id}/pre-email:
    post:
      operationId: generatePreEmail
      tags: [emails]
      summary: Generate a pre-sprint invitation email (AI-powered)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailGenerateInput"
      responses:
        "200":
          description: Generated pre-sprint email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailDraft"
  /sprints/{id}/post-email:
    post:
      operationId: generatePostEmail
      tags: [emails]
      summary: Generate a post-sprint analysis email (AI-powered)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailGenerateInput"
      responses:
        "200":
          description: Generated post-sprint email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailDraft"
...
components:
  schemas:
...
    Sprint:
      type: object
      properties:
        id:
          type: integer
        founderId:
          type: integer
        founderName:
          type: string
        companyName:
          type: string
        industry:
          type: ["string", "null"]
        stage:
          type: ["string", "null"]
        programName:
          type: ["string", "null"]
        partnerName:
          type: ["string", "null"]
        scheduledDate:
          type: string
        scheduledTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        totalDuration:
          type: ["string", "null"]
        consultantName:
          type: string
        sprintHost:
          type: ["string", "null"]
        coHost:
          type: ["string", "null"]
        status:
          type: string
          enum: [scheduled, completed, cancelled]
        tsheetUrl:
          type: ["string", "null"]
        fathomUrl:
          type: ["string", "null"]
        sprintNumber:
          type: ["integer", "null"]
        sessionType:
          type: ["string", "null"]
        paymentStatus:
          type: ["string", "null"]
        billedTo:
          type: ["string", "null"]
        billNumber:
          type: ["string", "null"]
        price:
          type: ["string", "number", "null"]
        week:
          type: ["integer", "null"]
        month:
          type: ["integer", "null"]
        cyYear:
          type: ["integer", "null"]
        fyYear:
          type: ["integer", "null"]
        quarter:
          type: ["string", "null"]
        strengths:
          type: ["string", "null"]
        gaps:
          type: ["string", "null"]
        swotAnalysis:
          type: ["string", "null"]
        nextGoal:
          type: ["string", "null"]
        actionableSteps:
          type: ["string", "null"]
        mentorshipRecommendation:
          type: ["string", "null"]
        marketConnections:
          type: ["string", "null"]
        meetLink:
          type: ["string", "null"]
        preEmailSentAt:
          type: ["string", "null"]
        postEmailSentAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - founderId
        - founderName
        - companyName
        - scheduledDate
        - consultantName
        - status
        - createdAt
    SprintInput:
      type: object
      properties:
        founderId:        { type: integer }
        scheduledDate:    { type: string }
        scheduledTime:    { type: string }
        endTime:          { type: string }
        consultantName:   { type: string }
        sprintHost:       { type: string }
        coHost:           { type: string }
        meetLink:         { type: string }
        tsheetUrl:        { type: string }
        fathomUrl:        { type: string }
        sprintNumber:     { type: integer }
        sessionType:      { type: string }
        paymentStatus:    { type: string }
        billedTo:         { type: string }
        billNumber:       { type: string }
        price:            { type: number }
      required:
        - founderId
        - scheduledDate
        - consultantName
    SprintUpdate:
      type: object
      properties:
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        consultantName:
          type: string
        status:
          type: string
          enum: [scheduled, completed, cancelled]
        tsheetUrl:
          type: string
        fathomUrl:
          type: string
        sprintNumber:
          type: integer
        sessionType:
          type: string
        strengths:
          type: string
        gaps:
          type: string
        swotAnalysis:
          type: string
        nextGoal:
          type: string
        actionableSteps:
          type: string
        mentorshipRecommendation:
          type: string
        marketConnections:
          type: string
        meetLink:
          type: string
    EmailGenerateInput:
      type: object
      properties:
        tone:
          type: string
          enum: [professional, friendly, concise, formal]
          default: professional
    EmailDraft:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
        to:
          type: string
        toName:
          type: string
        sprintId:
          type: integer
        emailType:
          type: string
          enum: [pre_sprint, post_sprint]
      required:
        - subject
        - body
        - to
        - toName
        - sprintId
        - emailType
...