# lib/api-zod/src/generated/types/sprint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Thinking Spree consultant platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SprintStatus } from "./sprintStatus";

export interface Sprint {
  id: number;
  founderId: number;
  founderName: string;
  companyName: string;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  stage?: string | null;
  /** @nullable */
  programName?: string | null;
  /** @nullable */
  partnerName?: string | null;
  scheduledDate: string;
  /** @nullable */
  scheduledTime?: string | null;
  /** @nullable */
  endTime?: string | null;
  /** @nullable */
  totalDuration?: string | null;
  consultantName: string;
  /** @nullable */
  sprintHost?: string | null;
  /** @nullable */
  coHost?: string | null;
  status: SprintStatus;
  /** @nullable */
  tsheetUrl?: string | null;
  /** @nullable */
  fathomUrl?: string | null;
  /** @nullable */
  sprintNumber?: number | null;
  /** @nullable */
  sessionType?: string | null;
  /** @nullable */
  paymentStatus?: string | null;
  /** @nullable */
  billedTo?: string | null;
  /** @nullable */
  billNumber?: string | null;
  /** @nullable */
  price?: string | number | null;
  /** @nullable */
  week?: number | null;
  /** @nullable */
  month?: number | null;
  /** @nullable */
  cyYear?: number | null;
  /** @nullable */
  fyYear?: number | null;
  /** @nullable */
  quarter?: string | null;
  /** @nullable */
  strengths?: string | null;
  /** @nullable */
  gaps?: string | null;
  /** @nullable */
  swotAnalysis?: string | null;
  /** @nullable */
  nextGoal?: string | null;
  /** @nullable */
  actionableSteps?: string | null;
  /** @nullable */
  mentorshipRecommendation?: string | null;
  /** @nullable */
  marketConnections?: string | null;
  /** @nullable */
  meetLink?: string | null;
  /** @nullable */
  preEmailSentAt?: string | null;
  /** @nullable */
  postEmailSentAt?: string | null;
  createdAt: string;
}